Move all deps to dependencies
