+++ DashAI/back/api/api_v1/endpoints/jobs.py
@@ -9,10 +9,10 @@ from sqlalchemy.orm import Session
 from DashAI.back.api.api_v1.schemas.job_params import JobParams
 from DashAI.back.containers import Container
 from DashAI.back.core.job_queue import job_queue_loop
+from DashAI.back.dependencies.job_queues import BaseJobQueue
+from DashAI.back.dependencies.job_queues.base_job_queue import JobQueueError
+from DashAI.back.dependencies.registry import ComponentRegistry
 from DashAI.back.job.base_job import BaseJob, JobError
-from DashAI.back.job_queues import BaseJobQueue
-from DashAI.back.job_queues.base_job_queue import JobQueueError
-from DashAI.back.services.registry import ComponentRegistry
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
