+++ DashAI/back/dependencies/database/__init__.py
@@ -1,0 +1,1 @@
+from DashAI.back.dependencies.database.sqlite_database import Base, SQLiteDatabase
