+++ DashAI/back/dependencies/registry/__init__.py
@@ -1,0 +1,3 @@
+from DashAI.back.dependencies.registry.component_registry import (  # noqa: F401
+    ComponentRegistry,
+)
