--- DashAI/back/job_queues/__init__.py
@@ -1,2 +1,0 @@
-from DashAI.back.job_queues.base_job_queue import BaseJobQueue
-from DashAI.back.job_queues.simple_job_queue import SimpleJobQueue
