--- DashAI/back/services/registry/__init__.py
@@ -1,3 +1,0 @@
-from DashAI.back.services.registry.component_registry import (  # noqa: F401
-    ComponentRegistry,
-)
