+++ tests/back/api/test_jobs.py
@@ -4,12 +4,12 @@ import joblib
 import pytest
 from fastapi.testclient import TestClient
 
-from DashAI.back.database.models import Experiment, Run
 from DashAI.back.dataloaders.classes.csv_dataloader import CSVDataLoader
+from DashAI.back.dependencies.database.models import Experiment, Run
+from DashAI.back.dependencies.registry import ComponentRegistry
 from DashAI.back.job.model_job import ModelJob
 from DashAI.back.metrics import BaseMetric
 from DashAI.back.models import BaseModel
-from DashAI.back.services.registry import ComponentRegistry
 from DashAI.back.tasks import BaseTask
