+++ tests/back/api/test_predict.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from fastapi.testclient import TestClient
 
-from DashAI.back.database.models import Experiment, Run
+from DashAI.back.dependencies.database.models import Experiment, Run
 
 
 @pytest.fixture(scope="module", name="dataset_id")
