+++ tests/back/api/test_runs_api.py
@@ -3,7 +3,7 @@ import json
 import pytest
 from fastapi.testclient import TestClient
 
-from DashAI.back.database.models import Dataset, Experiment
+from DashAI.back.dependencies.database.models import Dataset, Experiment
 
 
 @pytest.fixture(scope="module", name="experiment_id")
